requirements.txt
python-telegram-bot==20.7
yfinance==0.2.18
requests==2.31.0
httpx==0.25.0